changed from double to int

+++ include/matrix_generation.h
@@ -4,9 +4,9 @@
 #include "utils.h"
 
 // Generate a sparse matrix with given non-zero probability
-SparseMatrix* generate_sparse_matrix(double non_zero_prob);
+SparseMatrix* generate_sparse_matrix(int non_zero_prob);
 
 // Generate a pair of sparse matrices for multiplication
-void generate_matrix_pair(double non_zero_prob, SparseMatrix **X, SparseMatrix **Y);
+void generate_matrix_pair(int non_zero_prob, SparseMatrix **X, SparseMatrix **Y);
 
 #endif // MATRIX_GENERATION_H
